chore: Generate @fluentui/react-components peer dep safely
In order to guarantee that all new created packages will have a correct
minimum supported peer dependency, package.json is autogenerated with
the version installed in the current lockfile.

If authors need better backwards compatibility, they are responsible for
testing the earlier versions of react-components themselves and setting
the peer dependency manually.