fix: resolve Safari issue when referencing fallback
Uses the correct ID format when referencing a canvas
in `-webkit-canvas()`. Previously we used the CSS
ID selector (#) but we must omit the hash.

Safari uses an undocumented, vendor-specific API for
creating HTML canvases that can be efficiently rendered
as another element's background.

We use this API to polyfill-ish the CSS Custom Paint API
in Safari.